Defaults to Terraform 0.12.6 for tests
Defines the value of TERRAFORM_VERSION to the current version available
of the tool. Without this definition, the version downloaded is 0.11.11
as seen at `9070dcf` on `cloudposse/build-harness` repository.

Considering the TERRAFORM_VERSION declared on `build-harness` as the
upstream value of the in this repository, this default declaration
should be removed whenever possible. Meanwhile, if the variable is
already declared, it assumes the value declared to it.